Use uint64_t for string length in decode_string

+++ include/rive/core/reader.h
@@ -33,7 +33,7 @@ inline size_t decode_uint_leb(const uint8_t* buf, const uint8_t* buf_end, uint64
 /* Decodes a string
  */
 inline size_t
-decode_string(uint8_t str_len, const uint8_t* buf, const uint8_t* buf_end, char* char_buf) {
+decode_string(uint64_t str_len, const uint8_t* buf, const uint8_t* buf_end, char* char_buf) {
     // Return zero bytes read on buffer overflow
     if (buf_end - buf < str_len) {
         return 0;
